exercise-06 - JS modules init
