+++ index.html
@@ -9,6 +9,6 @@
 </head>
 <body>
 
-<script src='./scripts/main.js'></script>
+<script type="module" src='./scripts/main.js'></script>
 </body>
 </html>
