+++ scripts/utilis.js
@@ -1,0 +1,3 @@
+export function between(min, max) {
+    return Math.floor(Math.random() * (max - min + 1) + min);
+}
